Added web sockets to make the application more response when manipulating data.

+++ log_interface/routes/admin.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, session, jsonify
+from flask import Blueprint, session, jsonify, send_file
 from models.account_model import AccountModel
 
 """
@@ -18,14 +18,38 @@ def delete_account(account_id):
                 "status" : "error",
                 "message" : "Can't delete currently logged in account"
             }), 403
-        target_account = AccountModel.fetch_account_by_id(account_id)
-        target_account.delete_account()
-        return jsonify({
-            "status" : "ok",
-            "message" : account_id
-        }), 200
+        account = AccountModel.fetch_account_by_id(session["user_id"])
+        if account.is_admin():
+            target_account = AccountModel.fetch_account_by_id(account_id)
+            target_account.delete_account()
+            return jsonify({
+                "status" : "ok",
+                "message" : account_id
+            }), 200
+        else:
+            return jsonify({
+                "status" : "unauthorised",
+                "message" : account_id
+            }), 401
     except:
         return jsonify({
             "status" : "error",
             "message" : "Account not found"
         }), 404
+
+@admin_blueprint.route("/download_database", methods=['GET'])
+def download_database():
+    try:
+        account = AccountModel.fetch_account_by_id(session["user_id"])
+        if account.is_admin():
+            return send_file('../data/data.db', as_attachment=True)
+        else:
+            return jsonify({
+                "status" : "unauthorised",
+                "message" : session["user_id"]
+            }), 401
+    except FileNotFoundError:
+        return jsonify({
+                "status" : "unauthorised",
+                "message" : session["user_id"]
+            }), 404
